Added a new python project.
This project is about performing operations that can be performed using a scientific calculator which is coded in python. The scientific calculations that can be performed includes calculating percentage, factorial, LCM, HCF, power, exponential, logarithm and square root of a given number as many times as the user wants.
This is a simple project made using python basic topics such as datatypes and its methods, inbuilt libraries, iterations and its conditions.